+++ source/index.html
@@ -3,8 +3,10 @@
   <head>
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Интернет-магазин Кэт энерджи</title>
+    <link rel="preload" href="fonts/Oswald-Regular.woff2" as="font" crossorigin="anonymous">
+    <link rel="preload" href="fonts/Oswald-Medium.woff2" as="font" crossorigin="anonymous">
+    <link rel="preload" href="fonts/ArialMT.woff2" as="font" crossorigin="anonymous">
     <link rel="stylesheet" href="css/style.min.css">
   </head>
   <body>
@@ -63,13 +65,13 @@
               <h3 class="programs__title subtitle">Похудение</h3>
               <p class="programs__description">Ваш кот весит больше собаки и почти утратил способность лазить по деревьям?
                 Пора на диету! Cat Energy Slim поможет вашему питомцу сбросить лишний вес.</p>
-              <a class="programs__link" href="catalog.html">Каталог Slim</a>
+              <a class="programs__link programs__link--slim" href="catalog.html">Каталог Slim</a>
             </li>
             <li class="programs__item">
               <h3 class="programs__title subtitle">Набор массы</h3>
               <p class="programs__description">Заработать авторитет среди дворовых котов и&nbsp;даже собак?
                   Серия Cat Energy Pro поможет вашему коту нарастить необходимые мышцы!</p>
-              <a class="programs__link" href="catalog.html">Каталог pro</a>
+              <a class="programs__link programs__link--pro" href="catalog.html">Каталог pro</a>
             </li>
           </ul>
         </section>
@@ -117,23 +119,6 @@
             <div class="example__image-wrapper wrapper">
               <div class=" example__image example__image--before"></div>
               <div class="example__image example__image--after"></div>
-
-              <!-- <picture>
-                <source type="image/webp" media="(min-width: 1300px)" srcset="img/before-desktop1x.webp 1x, img/before-desktop2x.webp 2x">
-                <source type="image/webp" media="(min-width: 768px)" srcset="img/before-tablet1x.webp 1x, img/before-tablet2x.webp 2x">
-                <source type="image/webp" srcset="img/before-mobile1x.webp 1x, img/before-mobile2x.webp 2x">
-                <source media="(min-width: 1300px)" srcset="img/before-desktop1x.png 1x, img/before-desktop2x.png 2x">
-                <source media="(min-width: 768px)" srcset="img/before-tablet1x.png 1x, img/before-tablet2x.png 2x">
-                <img class="example__image-before" src="img/before-mobile1x.png" srcset="img/before-mobile2x.png 2x" width="320" height="240" alt="Фото до программы">
-              </picture>
-              <picture>
-                <source type="image/webp" media="(min-width: 1300px)" srcset="img/after-desktop1x.webp 1x, img/after-desktop2x.webp 2x">
-                <source type="image/webp" media="(min-width: 768px)" srcset="img/after-tablet1x.webp 1x, img/after-tablet2x.webp 2x">
-                <source type="image/webp" srcset="img/after-mobile1x.webp 1x, img/after-mobile2x.webp 2x">
-                <source media="(min-width: 1300px)" srcset="./img/after-desktop1x.png 1x, img/after-desktop2x.png 2x">
-                <source media="(min-width: 768px)" srcset="img/after-tablet1x.png 1x, img/after-tablet2x.png 2x">
-                <img class="example__image-after" src="img/after-mobile1x.png" srcset="img/after-mobile2x.png 2x" width="320" height="240" alt="Фото до программы">
-              </picture> -->
             </div>
             <div class="example__button-wrapper">
               <button class="example__before" type="button">Было</button>
@@ -169,7 +154,7 @@
             </svg>
         </div>
         <div class="page-footer__social social">
-          <ul class="social__list reset">
+          <ul class="social__list">
             <li class="social__item">
               <a class="social__link" href="#" aria-label="Мы в Вконтакте">
                 <svg class="social__icon social__icon--vkontakte" width="22" height="12">
@@ -204,13 +189,12 @@
       </div>
     </footer>
     <script src="js/navigation.js"></script>
-    <script src="js/picturefill.js"></script>
+    <script src="js/picturefill.min.js"></script>
     <script src="js/svg4everybody.min.js"></script>
     <script>
       svg4everybody(); // run it now or whenever you are ready
     </script>
-    <script src="js/map.js"></script>
+    <script src="js/script.min.js"></script>
     <script src="https://maps.googleapis.com/maps/api/js?key=AIzaSyCKSwPt7bJ8xrI76hkncI4fTTrC_2zTjJY&callback=initMap"></script>
-    <script src="js/slider.js"></script>
 </body>
 </html>
