+++ source/sass/blocks/advantages.scss
@@ -7,7 +7,7 @@
   display: inline-block;
   vertical-align: top;
 
-  margin-bottom: 36px;
+  margin-bottom: 35px;
 }
 
 .advantages__list {
@@ -171,14 +171,13 @@
     margin: 0;
 
     &::before {
-      content: "";
+      top: -2px;
+
       width: 80px;
       height: 80px;
     }
 
     &::after {
-      content: "";
-      position: absolute;
       top: 13px;
       left: 15px;
