+++ source/sass/blocks/contacts.scss
@@ -77,9 +77,9 @@
 
   .contacts__address {
     display: inline-block;
-    vertical-align: auto;
 
     margin-bottom: 19px;
+    width: 152px;
 
     text-align: right;
   }
@@ -113,7 +113,7 @@
     left: 40px;
 
     width: 570px;
-    padding: 59px 89px 57px 76px;
+    padding: 67px 89px 57px 76px;
 
     background: $white;
   }
