+++ source/sass/blocks/container.scss
@@ -1,10 +1,14 @@
 .container {
   padding-left: 20px;
   padding-right: 20px;
-  margin: auto;
+  margin: 0 auto;
 
 
   &--secondly {
+    padding: 0;
+  }
+
+  &--base {
     padding: 0;
   }
 }
@@ -22,7 +26,7 @@
     }
 
     &--thirdly {
-      padding-right: 63px;
+      padding-right: 66px;
       padding-left: 70px;
     }
   }
@@ -44,6 +48,10 @@
       max-width: 1300px;
       padding-left: 40px;
       padding-right: 40px;
+    }
+
+    &--base {
+      padding: 0;
     }
   }
 }
