+++ source/sass/blocks/products.scss
@@ -1,5 +1,5 @@
 .products {
-  padding-top: 17px;
+  padding-top: 18px;
 }
 
 .products__title {
@@ -174,6 +174,25 @@
   padding-bottom: 35px;
 }
 
+.products__section-wrapper {
+  position: relative;
+  padding-top: 24px;
+
+  &::after {
+    content: "";
+
+    z-index: -1;
+    position: absolute;
+    top: 43px;
+    right: 0;
+
+    width: 100vw;
+    height: 3px;
+
+    background: $green;
+  }
+}
+
 @media (min-width: $tablet-width) {
   .products {
     padding-top: 38px;
@@ -194,7 +213,7 @@
     position: relative;
 
     width: 324px;
-    margin-bottom: 141px;
+    margin-bottom: 139px;
     padding: 122px 38px 40px 38px;
 
     background: $grey;
@@ -254,7 +273,7 @@
 
     width: 100%;
     padding-top: 41px;
-    margin-bottom: 32px;
+    margin-bottom: 34px;
   }
 
   .products__caption {
@@ -291,12 +310,11 @@
   }
 
   .products__text {
-    margin: 0;
-    margin-bottom: 85px;
+    margin: 0 0 81px;
     width: 169px;
 
     font-size: 16px;
-    line-height: 16px;
+    line-height: 20px;
   }
 
   .products__intro {
@@ -309,26 +327,35 @@
   }
 
   .products__wrapper {
-    padding-top: 59px;
-    padding-bottom: 78px;
+    padding: 0;
+    padding-bottom: 77px;
   }
 
   .products__link {
     margin: 0;
     width: 100%;
   }
+
+  .products__section-wrapper {
+    padding-top: 58px;
+
+    &::after {
+      top: 89px;
+    }
+  }
 }
 
 @media (min-width: $desktop-width) {
   .products {
     padding-top: 57px;
+    margin-bottom: 67px;
   }
 
   .products__item {
     width: 245px;
     padding-bottom: 38px;
     padding-top: 124px;
-    margin-bottom: 144px;
+    margin-bottom: 141px;
 
     &:last-child {
       &::before {
@@ -342,6 +369,10 @@
     &:nth-last-child(-n+4) {
       margin-bottom: 0;
     }
+  }
+
+  .products__table {
+    margin-bottom: 32px;
   }
 
   .products__link-image {
@@ -358,7 +389,14 @@
     justify-content: space-between;
 
     width: 100%;
-    padding-top: 67px;
     padding-bottom: 86px;
+  }
+
+  .products__section-wrapper {
+    padding-top: 66px;
+
+    &::after {
+      top: 96px;
+    }
   }
 }
