+++ source/sass/blocks/promo.scss
@@ -18,7 +18,7 @@
 .promo__title {
   width: 250px;
   margin: 0;
-  margin-bottom: 22px;
+  margin-bottom: 21px;
 
   text-align: center;
