+++ source/sass/blocks/social.scss
@@ -3,13 +3,22 @@
   justify-content: space-between;
   align-items: center;
 
-  width: 137px;
+  width: 130px;
   margin: 0 auto;
+  list-style: none;
+  padding: 0;
 }
 
 .social__item {
   width: 22px;
   align-self: center;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.social__link {
+  display: inline-block;
 }
 
 .social__icon {
@@ -47,6 +56,7 @@
 
   .social__item {
     width: 29px;
+    justify-content: stretch;
   }
 
   .social__icon {
