+++ source/sass/fonts.scss
@@ -4,6 +4,7 @@
   src: url("../fonts/Oswald-Medium.woff") format("woff");
   font-weight: 500;
   font-style: normal;
+  font-display: swap;
 }
 
 @font-face {
@@ -12,6 +13,7 @@
   src: url("../fonts/Oswald-Regular.woff") format("woff");
   font-weight: normal;
   font-style: normal;
+  font-display: swap;
 }
 
 @font-face {
@@ -20,4 +22,5 @@
   src: url("../fonts/ArialMT.woff") format("woff");
   font-weight: 500;
   font-style: normal;
+  font-display: swap;
 }
